crowd density estimation/flaskapp/requirements.txt
matplotlib==3.6.0
numpy==1.23.3
torch==1.12.1
torchvision==0.13.1
flask==2.2.2
requests==2.28.1
Pillow==9.2.0
waitress
opencv-python==4.6.0
